Extract a DynamicFilter component
