+++ src/components/Filters/DynamicSourceFilters.js
@@ -5,16 +5,9 @@ import Drawer from "@material-ui/core/Drawer";
 import FormGroup from "@material-ui/core/FormGroup";
 import FilterActions from "components/Filters/FilterActions";
 import { makeStyles } from "@material-ui/styles";
-import {
-  selectFiltersLength,
-  selectFilterTypeAtIndex
-} from "redux-ducks/filters";
+import { selectFiltersLength } from "redux-ducks/filters";
 import { useSelector } from "react-redux";
-import FilterTextField from "components/Filters/FilterTextField";
-import FilterSelect from "components/Filters/FilterSelect";
-import FilterSort from "components/Filters/FilterSort";
-import FilterTristate from "components/Filters/FilterTristate";
-import FilterGroup from "components/Filters/FilterGroup";
+import DynamicFilter from "components/Filters/DynamicFilter";
 import times from "lodash/times";
 
 // FIXME: Weird blue line when clicking the <FormGroup>
@@ -90,36 +83,6 @@ const DynamicSourceFilters = () => {
       </Drawer>
     </React.Fragment>
   );
-};
-
-type Props = { index: number };
-
-const DynamicFilter = ({ index }: Props) => {
-  const type = useSelector(state => selectFilterTypeAtIndex(state, index));
-
-  if (["HEADER", "SEPARATOR", "CHECKBOX"].includes(type)) {
-    console.error(`Catalogue filters - ${type} is not implemented.`);
-  }
-
-  switch (type) {
-    case "TEXT":
-      return <FilterTextField index={index} key={index} />;
-
-    case "SELECT":
-      return <FilterSelect index={index} key={index} />;
-
-    case "SORT":
-      return <FilterSort index={index} key={index} />;
-
-    case "TRISTATE":
-      return <FilterTristate index={index} key={index} />;
-
-    case "GROUP":
-      return <FilterGroup index={index} key={index} />;
-
-    default:
-      return null;
-  }
 };
 
 export default DynamicSourceFilters;
